feat: refactor useEditingState hook with type definitions - Extract type definitions to separate types.ts file - Improve hook structure and maintainability - Add proper TypeScript interfaces
