+++ frontend/src/hooks/useEditingState/types.ts
@@ -1,0 +1,6 @@
+export interface EditingStateOperations<T> {
+  editingItem: T | null;
+  isEditing: boolean;
+  startEditing: (item: T) => void;
+  stopEditing: () => void;
+}
